Changed testing to use grpctest, changed t.Errorf to t.Fatalf, moved helper function to util.go
